% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparaConfig.R
\name{cat_ParaValue}
\alias{cat_ParaValue}
\title{cat_ParaValue}
\usage{
cat_ParaValue(parameter, values)
}
\arguments{
\item{parameter}{the address to the parameter that needs to be modified}

\item{values}{a string of values that are separated by comma}
}
\value{
Character string. The string is the path of apsimx submodule and
  value pair
}
\description{
A wrapper function of `paste` to construct configuration values
  for `ApsimX` Edit feature.
}
\details{
when replace the number of values must match the number of layers
soil parameters.
}
