% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUL_LL.R
\name{doDUL_LL_range}
\alias{doDUL_LL_range}
\title{doDUL_LL_range}
\usage{
doDUL_LL_range(
  SW,
  id.vars = id_vars,
  startd = "2011-01-01",
  endd = "2012-06-30"
)
}
\arguments{
\item{SW}{data.table. Mean soil water content measurments with `Experiment`,
`SowingDate` and `Depth` as three key columns.}

\item{endd}{}
}
\value{
Max and Min VWC
}
\description{
A wrapper function to be used in data.table syntax with lappy.
  `DUL_LL` function will find the maximum and minimum water content (mm) in a
  given layer of soil and return its volumetric water content.
}
