% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUL_LL.R
\name{DUL_LL}
\alias{DUL_LL}
\title{DUL_LL}
\usage{
DUL_LL(x, thickness)
}
\arguments{
\item{x}{A numeric vector. Soil water}

\item{thickness}{Integer. The depth of a given soil layer in mm.}
}
\value{
Max and Min VWC
}
\description{
A wrapper function to be used in data.table syntax with lappy.
  `DUL_LL` function will find the maximum and minimum water content (mm) in a
  given layer of soil and return its volumn metric content.
}
