% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditApsimx.R
\name{EditLayerKL}
\alias{EditLayerKL}
\title{EditLayerKL}
\usage{
EditLayerKL(layer, value, path, apsimx, saveTo)
}
\arguments{
\item{layer}{an integer for which layer needs to be modified}

\item{value}{a floating point number. must be less than 1. this value should
be sesenable biologically}

\item{path}{the path to apsimx `Models.exe`}

\item{apsimx}{the path to the base `.apsimx` file}

\item{saveTo}{a path to show where to save modified file. This is for
preventing the **Edit** flag overwriten the base `.apsimx` file.}
}
\value{

}
\description{
Edit a base `.apsimx` file with a _kl_ value to update a
  specific layer in the soil model
}
