% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_stats.R
\name{extract_trts}
\alias{extract_trts}
\title{extract_trts}
\usage{
extract_trts(
  filename,
  pattern = "(Ashley|Iversen).+SD\\\\d{1,2}",
  pattern_split = "(.+)(SD\\\\d{1,2})"
)
}
\arguments{
\item{filename}{a string or a full path of the db name}

\item{pattern}{a regex pattern to remove extra strings other than treatment
information}
}
\value{
a string with two elements. first is the site and the second is the
  sowing date
}
\description{
extract treatment information from output db names by using
  regular expression. It currently only works for two treatments.
}
