% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUL_LL.R
\name{filter_datemax}
\alias{filter_datemax}
\title{filter_datemax}
\usage{
filter_datemax(mean_SW, mode = c("max", "min"), id.vars = id_vars)
}
\arguments{
\item{mean_SW}{data.table has mean soil water measurements.}

\item{mode}{character string to indicate which mode to filter: "max" or "min"}

\item{id.vars}{a character vector indicates the grouping variables}
}
\value{

}
\description{
need to know which date has the maximum and minimum SW to figure
  out the range from the replicates
}
