% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{group_in_season}
\alias{group_in_season}
\title{group_in_season}
\usage{
group_in_season(DT)
}
\arguments{
\item{DT, }{data.table}
}
\value{
data.table with a column `Season`
}
\description{
Group annual data into seasonal data 1 July to 30 June next year
}
\details{
the data.table must have a column named as "Date"
}
