% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_stats.R
\name{manipulate}
\alias{manipulate}
\title{manipulate}
\usage{
manipulate(DT_obs = obs, subset_cols = NULL, DT_pred = dt)
}
\arguments{
\item{DT_obs}{a data table that subsetted from the aggregated raw data}

\item{subset_cols}{a string vector. The column names that unwanted in the
prediction table. The default mode will remove "CheckpointID",
"Experiment", "FolderName"}

\item{DT_pred}{a data table that has the simulation results}
}
\value{
a data table has the same number of rows as the observation
  data.table
}
\description{
Take a subset data.table subsetted by the treament and join it
  with the corresponding simulation results
}
