% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_paras.R
\name{plot_params}
\alias{plot_params}
\title{plot_params}
\usage{
plot_params(
  DT = top5,
  title,
  format = "pdf",
  col_pred = "PSWC",
  col_obs = "SWC",
  point_size = 2,
  Depth = NULL,
  width = 10,
  height = 10
)
}
\arguments{
\item{Depth}{default is profile. can be set the "Depth" to check each layer.}

\item{height}{}
}
\value{

}
\description{

}
