% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_paras.R
\name{plot_root}
\alias{plot_root}
\title{Title}
\usage{
plot_root(
  DT = top1,
  title,
  format = "pdf",
  col_pred = "PSWC",
  col_obs = "SWC",
  point_size = 2,
  Depth = NULL,
  width = 10,
  height = 10
)
}
\arguments{
\item{DT}{data.table that has predictio and observation data with stats}

\item{height}{}
}
\value{

}
\description{
Title
}
