% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_stats.R
\name{read_dbtab}
\alias{read_dbtab}
\title{read_dbtab}
\usage{
read_dbtab(path = "./03processed-data/Richard.sqlite3",
  table = "SoilWater")
}
\arguments{
\item{path}{the path to the sqlite file}

\item{table}{a charater string for the table name}
}
\value{
A data.table has simulation results.
}
\description{
Connet to `sqlite3` file and read a specific table and convert
  the date string to a date type
}
