% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMet.R
\name{read_met_col}
\alias{read_met_col}
\title{read_met_col}
\usage{
read_met_col(path = path_met, skip = 7)
}
\arguments{
\item{path}{The path to access the met files.}

\item{skip}{The lines to skip for reading the header.}
}
\value{
A data .table and .frame is returned, which has only the colnames for met files.
}
\description{
A helper function to assist read_met.
It only read the colnames.
}
\examples{
\dontrun{
read_met_col("path", skip = 7)
}
}
\concept{read_met}
