% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_stats.R
\name{sims_stats}
\alias{sims_stats}
\title{sims_stats}
\usage{
sims_stats(
  pred_obs,
  keys = c("Experiment", "SimulationID", "SowingDate", "KLR", "RFV", "SKL"),
  col_pred = "PSWC",
  col_obs = "SWC"
)
}
\arguments{
\item{pred_obs}{a data.table which merged from the simulation table and
observation data.table}

\item{keys}{grouping variables}

\item{col_pred}{the colname for prediction result}

\item{col_obs}{the colname for observation}
}
\value{

}
\description{
Process one kl set with six KLR and six RFV
}
