% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_paras.R
\name{subset_obs}
\alias{subset_obs}
\title{subset_obs}
\usage{
subset_obs(
  DT = SW_mean,
  col_treatment1 = "Experiment",
  col_treatment2 = "SowingDate",
  treatment1,
  treatment2
)
}
\arguments{
\item{treatment2}{}
}
\value{

}
\description{
subset the obersvation data based on treatment
}
