% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_paras.R
\name{subset_pred}
\alias{subset_pred}
\title{subset_pred}
\usage{
subset_pred(
  DT = top5,
  col_treatment1 = "Experiment",
  col_treatment2 = "SowingDate",
  treatment1,
  treatment2
)
}
\arguments{
\item{treatment2}{}
}
\value{

}
\description{
subset_pred
}
