% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_paras.R
\name{subset_stats}
\alias{subset_stats}
\title{subset_stats}
\usage{
subset_stats(DT = DT_stats_layerKL)
}
\arguments{
\item{DT}{A data.table with two nested list columns: `data` and `stats`.
Column names `Source, Experiment, SowingDate, Depth` must be in the DT as
group variables}
}
\value{
a data.table has the best fit parameter for each layer in each
  treatment.
}
\description{
Unlist the stats result to obtain best fit for each layer and
  join back to the data for further visulisation
}
\examples{
subset_stats(DT = DT_stats_layerKL)
}
