% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_paras.R
\name{subsetByTreatment}
\alias{subsetByTreatment}
\title{subsetByTreatment}
\usage{
subsetByTreatment(DT = SW_mean, mode = NULL,
  col_treatment1 = "Experiment", col_treatment2 = "SowingDate",
  treatment1 = "Iversen12", treatment2 = "SD1")
}
\arguments{
\item{DT}{Data.table. The data.table contains observation or prediction
values for all treatments.}

\item{mode}{Default NULL is subset}

\item{col_treatment1}{Character string. The column name for treatment 1}

\item{col_treatment2}{Character string. The column name for treatment 2}

\item{treatment1}{The treatment value that wants to be subset.}

\item{treatment2}{The treatment value that wants to be subset.}
}
\value{
A filtered data.table.
}
\description{
Subset the observation/prediction data based on treatment.
  Current version can only do two treatment subset.
}
\examples{
subsetByTreatment(DT = SW_mean,
                  col_treatment1 = "Experiment",
                  col_treatment2 = "SowingDate",
                  treatment1 = "Iversen12",
                  treatment2 = "SD1")
}
