% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparaConfig.R
\name{cat_ParaValue}
\alias{cat_ParaValue}
\title{cat_ParaValue}
\usage{
cat_ParaValue(parameter, values)
}
\arguments{
\item{parameter}{the address to the parameter that needs to be modified}

\item{values}{a string of values that are separated by comma}
}
\description{
cat_ParaValue
}
\details{
when replace the number of values must match the number of layers
soil parameters.
}
