% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMet.R
\name{read_met}
\alias{read_met}
\title{read_met}
\usage{
read_met(
  path = path_met,
  skip_unit = 9,
  skip_meta = 7,
  startd = "2010-10-01",
  endd = "2012-08-01",
  site = "AshleyDene"
)
}
\arguments{
\item{path}{A character string. The path to access the met files.}

\item{skip_unit}{An integer. The number of rows for skipping the unit line in met files.}

\item{skip_meta}{An integer. The number of rows for skipping the meta data before the column names start.}

\item{site}{}
}
\value{
A data .table and .frame is returned.
}
\description{
read_met
}
\examples{
\dontrun{
read_met("path", skip_unit = 9, skip_meta = 7)
}
}
