% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_stats.R
\name{sims_stats_multi}
\alias{sims_stats_multi}
\title{sims_stats_multi}
\usage{
sims_stats_multi(
  path_sims,
  pattern = ".db$",
  DT_observation,
  tableName = "Report",
  pattern_trts = "(Ashley|Iversen).+SD\\\\d{1,2}",
  pattern_split = "(.+)(SD\\\\d{1,2})",
  col_treatment1 = "Experiment",
  col_treatment2 = "SowingDate",
  mode = c("Profile", "Layers", "Manual"),
  keys = c("Experiment", "SimulationID", "SowingDate", "KLR", "RFV", "SKL")
)
}
\arguments{
\item{path_sims}{The path to all apsimx output}

\item{pattern}{db extension}

\item{DT_observation}{The data.table has aggregated observations and ordered
by days}

\item{tableName}{a charater string. The name of the simulation output table
in the db}

\item{pattern_trts}{regular expression to extract the treatment names from
filename}

\item{pattern_split}{regular expression to split two treatment names}

\item{col_treatment1}{Default is the "Experiment"}

\item{col_treatment2}{Default is the "SowingDate"}

\item{mode}{three options:
\itemize{
  \item _Profile_ will calculate the soil water profile statisic.
  \item _Layers_ will calculate the stats for each layer.
  \item _Manual_ will calculate the layer by layer optimisation which will be
  the results from simulations in all combinations of all layer and possible
  kl values.
}}

\item{keys}{a character vector to define the keys for grouping. three options
  to match the mode:
\itemize{
  \item _Profile_ mode - The default keys are _"Experiment", "SimulationID", "SowingDate","KLR","RFV","SKL"_ for
  soil water profile.
  \item _Layers_ mode - _"Depth"_ should be added into the default keys.
  \item _Manual_ mode - _"Experiment", "SowingDate", "Depth"_ are the keys.
  }}
}
\value{
a list
}
\description{
assemble the function above to process multiple dbs
}
