% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_stats.R
\name{sims_stats_multi}
\alias{sims_stats_multi}
\title{sims_stats_multi}
\usage{
sims_stats_multi(path_sims, pattern = ".db$", DT_observation)
}
\arguments{
\item{pattern}{db extension}

\item{DT_observation}{}
}
\value{

}
\description{
assemble the function above to process multiple dbs
}
